bim/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
croniter==6.0.0
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.6
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
numpy==2.3.2
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.2
phonenumbers==9.0.13
pillow==11.3.0
procrastinate==3.5.2
psycopg==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pyotp==2.9.0
pypdf==6.0.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.3
requests==2.32.5
RouterOS-api==0.21.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
